add showQuickPick() and withProgress() to VscePromptApi
